Add modbus_reply_callback() and modbus_set_reply_callback()
Based on @frodete's proposal (#319) this implementation adds a
a callback-based reply-functionality.

The old mapping-implementation is now based on this version. As a nice
side-effect some DRY code optimizations could have been done.

Also adds a MODBUS_SLAVE_ACCEPT_ALL "address", which can be used with
modbus_set_slave(). This makes the modbus_receive()-function accept
all requests, independently of the slave-address found in the
decoded-request. Useful for multi-slave-implementation using libmodbus.